Implement segment references
Expose the references of a segment via a child node 'references'. Each
child of 'references' is a segment node, whose name is the index of that
segment in the collection of references.

Example:

    /store/<t>/<s>/references/0/...
    /store/<t>/<s>/references/1/...

where <t> is the name of a TAR file and <s> is the ID of a segment in
<t>.