Implement segment records
Expose the records in a segment via a child node named 'records'. Each
child of 'records' represent a record in that segment. The name of these
children is the record number.

Example:

    /store/<t>/<s>/records/0/...
    /store/<t>/<s>/records/1/...

where <t> is the name of a TAR file and <s> is the ID of a segment
inside <t>.

Each record node exposes the number, the offset and the type of the
corresponding record.